App/web/requirements.txt
asgiref==3.7.2
azure-core==1.29.5
azure-identity==1.15.0
azure-servicebus==7.11.4
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
Flask==3.0.0
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msal==1.26.0
msal-extensions==1.0.0
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
